+++ context/.storybook/main.js
@@ -1,28 +1,26 @@
-const path = require("path");
-const {
-  alias
-} = require('../build-utils/alias');
+const path = require('path');
+const { alias } = require('../build-utils/alias');
 module.exports = {
-  "stories": ['../app/static/js/shared-styles/**/*.stories.js', '../app/static/js/components/**/*.stories.js'],
-  "addons": ["@storybook/addon-links", "@storybook/addon-essentials", 'storybook-addon-swc'],
-  "webpackFinal": async config => {
+  stories: ['../app/static/js/**/*.stories.@(js|jsx|ts|tsx)'],
+  addons: ['@storybook/addon-links', '@storybook/addon-essentials', 'storybook-addon-swc'],
+  webpackFinal: async (config) => {
     config.resolve.alias = {
       // extend alias with our own
       ...config.resolve?.alias,
-      ...alias
+      ...alias,
     };
     config.module.rules.push({
       test: /\.mjs$/,
       include: /node_modules/,
-      type: "javascript/auto"
+      type: 'javascript/auto',
     });
     return config;
   },
   framework: {
-    name: "@storybook/react-webpack5",
-    options: {}
+    name: '@storybook/react-webpack5',
+    options: {},
   },
   docs: {
-    autodocs: true
-  }
+    autodocs: true,
+  },
 };