+++ context/app/static/js/components/Contexts.tsx
@@ -2,8 +2,24 @@ import { useContext, createContext } from 'js/helpers/context';
 
 // TODO: Continue populating these types as we find more of the uses of the flask data and app contexts
 
+type DagProvenanceType =
+  | {
+      origin: string;
+    }
+  | {
+      name: string;
+    };
+
 type FlaskDataContextType = {
   redirected_from: string;
+  entity: {
+    entity_type: string;
+    metadata: {
+      dag_provenance_list: DagProvenanceType[];
+      [key: string]: unknown;
+    };
+    [key: string]: unknown;
+  };
   [key: string]: unknown;
 };