+++ context/app/static/js/components/Header/AtlasToolsLinks/index.js
@@ -1,3 +1,0 @@
-import AtlasToolsLinks from './AtlasToolsLinks';
-
-export default AtlasToolsLinks;