+++ context/app/static/js/components/detailPage/files/FilesConditionalLink/index.js
@@ -1,3 +1,0 @@
-import FilesConditionalLink from './FilesConditionalLink';
-
-export default FilesConditionalLink;